DegradNet-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
torch==2.1.2
torchvision==0.16.2
Pillow==10.2.0
numpy==1.26.3
matplotlib==3.8.2
python-multipart==0.0.6
opencv-python-headless
segmentation-models-pytorch
